Respond to more PR #29 comments
